tests: fix signed overflow issues
* tests/test-dynarray.c (value_at): Avoid undefined behavior
in signed integer multiplication.
* tests/test-scratch-buffer.c (byte_at): Likewise, for the
theoretically-possible case where size_t is narrower than int.

+++ tests/test-dynarray.c
@@ -28,7 +28,7 @@
 #define N 100000
 
 static int
-value_at (int i)
+value_at (long long int i)
 {
   return (i % 13) + ((i * i) % 251);
 }
